+++ src/main/java/bridge/domain/bridge/Bridge.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.domain.bridge;
 
 import java.util.List;
