+++ src/main/java/bridge/domain/bridge/BridgeSize.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.domain.bridge;
 
 import bridge.domain.exception.WrongRangeSizeException;
