+++ src/main/java/bridge/domain/result/BridgeGameResult.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.domain.result;
 
 public class BridgeGameResult {
