+++ src/test/java/bridge/domain/GameCommandTest.java
@@ -5,7 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import bridge.Application;
-import bridge.utils.CommandInputParser;
+import bridge.domain.command.GameCommand;
+import bridge.utils.parser.CommandInputParser;
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
