+++ src/test/java/bridge/domain/MovingCommandTest.java
@@ -3,7 +3,8 @@ package bridge.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import bridge.utils.CommandInputParser;
+import bridge.domain.command.MovingCommand;
+import bridge.utils.parser.CommandInputParser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
